+++ projects/api/src/contracts/_internal/request/pageQuerySchema.ts
@@ -1,10 +1,10 @@
 import { z } from '../z.ts';
 
 export const pageQuerySchema = z.object({
-  page: z.number().optional().openapi({
+  page: z.number().int().optional().openapi({
     description: 'The page number to retrieve',
   }),
-  limit: z.number().optional().openapi({
+  limit: z.number().int().optional().openapi({
     description: 'The number of items per page',
   }),
 });
