+++ projects/api/src/contracts/_internal/request/watchNowParamsSchema.ts
@@ -2,7 +2,7 @@ import { z } from '../z.ts';
 import { countryParamsSchema } from './countryParamsSchema.ts';
 
 export const watchNowParamsSchema = z.object({
-  id: z.string().or(z.number()).openapi({
+  id: z.string().or(z.number().int()).openapi({
     description:
       'The Trakt ID of the resource to get the watch now sources of.',
   }),
