+++ projects/api/src/contracts/_internal/response/episodeIdsResponseSchema.ts
@@ -1,8 +1,8 @@
 import { z } from '../z.ts';
 
 export const episodeIdsResponseSchema = z.object({
-  trakt: z.number(),
-  tvdb: z.number(),
+  trakt: z.number().int(),
+  tvdb: z.number().int(),
   imdb: z.string().nullable(),
-  tmdb: z.number(),
+  tmdb: z.number().int(),
 });
