+++ projects/api/src/contracts/_internal/response/episodeStatsResponseSchema.ts
@@ -1,10 +1,10 @@
 import { z } from '../z.ts';
 
 export const episodeStatsResponseSchema = z.object({
-  watchers: z.number(),
-  plays: z.number(),
-  collectors: z.number(),
-  comments: z.number(),
-  lists: z.number(),
-  votes: z.number(),
+  watchers: z.number().int(),
+  plays: z.number().int(),
+  collectors: z.number().int(),
+  comments: z.number().int(),
+  lists: z.number().int(),
+  votes: z.number().int(),
 });
