+++ projects/api/src/contracts/_internal/response/mediaMutationCountsSchema.ts
@@ -5,8 +5,8 @@ import { z } from '../z.ts';
  * where only subsets are returned
  */
 export const mediaMutationCountsSchema = z.object({
-  movies: z.number(),
-  shows: z.number(),
-  seasons: z.number(),
-  episodes: z.number(),
+  movies: z.number().int(),
+  shows: z.number().int(),
+  seasons: z.number().int(),
+  episodes: z.number().int(),
 });
