+++ projects/api/src/contracts/_internal/response/movieIdsResponseSchema.ts
@@ -1,8 +1,8 @@
 import { z } from '../z.ts';
 
 export const movieIdsResponseSchema = z.object({
-  trakt: z.number(),
+  trakt: z.number().int(),
   slug: z.string(),
   imdb: z.string(),
-  tmdb: z.number(),
+  tmdb: z.number().int(),
 });
