+++ projects/api/src/contracts/_internal/response/movieStatsResponseSchema.ts
@@ -2,6 +2,6 @@ import { z } from '../z.ts';
 import { episodeStatsResponseSchema } from './episodeStatsResponseSchema.ts';
 
 export const movieStatsResponseSchema = episodeStatsResponseSchema.extend({
-  favorited: z.number(),
-  recommended: z.number(),
+  favorited: z.number().int(),
+  recommended: z.number().int(),
 });
