+++ projects/api/src/contracts/_internal/response/seasonIdsResponseSchema.ts
@@ -1,7 +1,7 @@
 import { z } from '../z.ts';
 
 export const seasonIdsResponseSchema = z.object({
-  trakt: z.number(),
-  tvdb: z.number().nullable(),
-  tmdb: z.number(),
+  trakt: z.number().int(),
+  tvdb: z.number().int().nullable(),
+  tmdb: z.number().int(),
 });
