+++ projects/api/src/contracts/_internal/response/showStatsResponseSchema.ts
@@ -2,5 +2,5 @@ import { z } from '../z.ts';
 import { movieStatsResponseSchema } from './movieStatsResponseSchema.ts';
 
 export const showStatsResponseSchema = movieStatsResponseSchema.extend({
-  collected_episodes: z.number(),
+  collected_episodes: z.number().int(),
 });
