+++ projects/api/src/contracts/_internal/response/statsResponseSchema.ts
@@ -1,7 +1,7 @@
 import { z } from '../z.ts';
 
 export const statsResponseSchema = z.object({
-  play_count: z.number(),
-  minutes_watched: z.number(),
-  minutes_left: z.number().optional(),
+  play_count: z.number().int(),
+  minutes_watched: z.number().int(),
+  minutes_left: z.number().int().optional(),
 });
