+++ projects/api/src/contracts/_internal/response/studioResponseSchema.ts
@@ -5,7 +5,7 @@ export const studioResponseSchema = z.object({
   country: z.string().optional(),
   ids: z.object({
     slug: z.string(),
-    trakt: z.number(),
-    tmdb: z.number().optional(),
+    trakt: z.number().int(),
+    tmdb: z.number().int().optional(),
   }),
 });
