+++ projects/api/src/contracts/_internal/response/watchedStatsResponseSchema.ts
@@ -1,7 +1,7 @@
 import { z } from '../z.ts';
 
 export const watchedStatsResponseSchema = z.object({
-  watcher_count: z.number(),
-  play_count: z.number(),
-  collected_count: z.number(),
+  watcher_count: z.number().int(),
+  play_count: z.number().int(),
+  collected_count: z.number().int(),
 });
