+++ projects/api/src/contracts/checkin/_internal/request/showCheckinRequestSchema.ts
@@ -5,10 +5,10 @@ import { sharingRequestSchema } from './sharingRequestSchema.ts';
 const showAbsoluteCheckinRequestSchema = z.object({
   show: z.object({
     title: z.string().optional(),
-    year: z.number().optional(),
+    year: z.number().int().optional(),
     ids: showIdsRequestSchema,
   }),
-  episode: z.object({ number_abs: z.number() }),
+  episode: z.object({ number_abs: z.number().int() }),
   sharing: sharingRequestSchema,
   message: z.string().optional(),
 });
@@ -22,10 +22,10 @@ const episodeCheckinRequestSchema = z.object({
 const episodeWithShowCheckinRequestSchema = z.object({
   show: z.object({
     title: z.string().optional(),
-    year: z.number().optional(),
+    year: z.number().int().optional(),
     ids: showIdsRequestSchema,
   }),
-  episode: z.object({ season: z.number(), number: z.number() }),
+  episode: z.object({ season: z.number().int(), number: z.number().int() }),
   sharing: sharingRequestSchema,
   message: z.string().optional(),
 });
