+++ projects/api/src/contracts/oauth/_internal/response/codeResponseSchema.ts
@@ -4,6 +4,6 @@ export const codeResponseSchema = z.object({
   device_code: z.string(),
   user_code: z.string(),
   verification_url: z.string(),
-  expires_in: z.number(),
-  interval: z.number(),
+  expires_in: z.number().int(),
+  interval: z.number().int(),
 });
