+++ projects/api/src/contracts/recommendations/_internal/request/recommendationsQuerySchema.ts
@@ -2,10 +2,10 @@ import { ignoreQuerySchema } from '../../../_internal/request/ignoreQuerySchema.
 import { z } from '../../../_internal/z.ts';
 
 export const recommendationsQuerySchema = z.object({
-  limit: z.number().openapi({
+  limit: z.number().int().openapi({
     description: 'Limit the number of results.',
   }),
-  watch_window: z.number().optional().openapi({
+  watch_window: z.number().int().optional().openapi({
     description: 'The watch window in days for the recommendations.',
   }),
 }).merge(ignoreQuerySchema);
