+++ projects/api/src/contracts/search/_internal/response/searchMovieResponseSchema.ts
@@ -3,6 +3,6 @@ import { z } from '../../../_internal/z.ts';
 
 export const searchMovieResponseSchema = z.object({
   type: z.literal('movie'),
-  score: z.number(),
+  score: z.number().int(),
   movie: movieResponseSchema,
 });
