+++ projects/api/src/contracts/shows/_internal/response/showStreamingResponseSchema.ts
@@ -2,7 +2,7 @@ import { showResponseSchema } from '../../../_internal/response/showResponseSche
 import { z } from '../../../_internal/z.ts';
 
 export const showStreamingResponseSchema = z.object({
-  rank: z.number(),
-  delta: z.number().nullable(),
+  rank: z.number().int(),
+  delta: z.number().int().nullable(),
   show: showResponseSchema,
 });
