+++ projects/api/src/contracts/sync/_internal/response/historyResponseSchema.ts
@@ -2,7 +2,7 @@ import { bulkMediaRequestSchema } from '../../../_internal/request/bulkMediaRequ
 import { z } from '../../../_internal/z.ts';
 
 export const historyResponseSchema = z.object({
-  added: z.object({ movies: z.number(), episodes: z.number() }),
-  updated: z.object({ movies: z.number(), episodes: z.number() }),
+  added: z.object({ movies: z.number().int(), episodes: z.number().int() }),
+  updated: z.object({ movies: z.number().int(), episodes: z.number().int() }),
   not_found: bulkMediaRequestSchema,
 });
