+++ projects/api/src/contracts/users/_internal/request/reorderRequestSchema.ts
@@ -1,5 +1,5 @@
 import { z } from '../../../_internal/z.ts';
 
 export const reorderRequestSchema = z.object({
-  rank: z.array(z.number()),
+  rank: z.array(z.number().int()),
 });
