+++ projects/api/src/contracts/users/_internal/response/hiddenAddResponseSchema.ts
@@ -3,9 +3,9 @@ import { z } from '../../../_internal/z.ts';
 
 export const hiddenAddResponseSchema = z.object({
   added: z.object({
-    movies: z.number(),
-    shows: z.number(),
-    season: z.number(),
+    movies: z.number().int(),
+    shows: z.number().int(),
+    season: z.number().int(),
   }),
   not_found: bulkMediaRequestSchema,
 });
