+++ projects/api/src/contracts/users/_internal/response/reorderListsResponseSchema.ts
@@ -1,6 +1,6 @@
 import { z } from '../../../_internal/z.ts';
 
 export const reorderListsResponseSchema = z.object({
-  updated: z.number(),
-  skipped_ids: z.array(z.number()),
+  updated: z.number().int(),
+  skipped_ids: z.array(z.number().int()),
 });
